ref: Updated router class
