+++ src/constants.ts
@@ -1,0 +1,17 @@
+const HTTP_METHODS = {
+  GET: 'GET',
+  POST: 'POST',
+  PUT: 'PUT',
+  DELETE: 'DELETE',
+  PATCH: 'PATCH',
+  HEAD: 'HEAD',
+  OPTIONS: 'OPTIONS',
+  CONNECT: 'CONNECT',
+  TRACE: 'TRACE',
+};
+
+const REGEX_FOR_ID = new RegExp(
+  '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$',
+);
+
+export { HTTP_METHODS, REGEX_FOR_ID };
